dlchess/requirements.txt
tensorflow
numpy
keras
python-chess